get_group {neon}	R Documentation

get_group

Description

Returns a dataframe of the groups accessible by your Smartabase account. For more details see the help vignette: vignette("getting-started", package = "neon")

Usage

get_group(url = NULL, username = NULL, password = NULL, id_message = FALSE)

Arguments

url
string: Smartabase url e.g. 'example.smartabase.com/site' – ignore if setup with save_credentials()

username
string: Smartabase username – ignore if setup with save_credentials()

password
string: Smartabase password – ignore if setup with save_credentials()

id_message
boolean: if TRUE, displays messages to screen about retrieving user ID information. Only relevant when get_id = TRUE

Value

dataframe of available groups

Examples

## Not run: 
# Assuming credentials have been setup with save_credentials(), otherwise
# must supply username, password and url:
get_group()

## End(Not run)

[Package neon version 0.1.20 Index]