get_metadata_names {neon}	R Documentation

get_metadata_names

Description

Returns vector of metadata variables present in data frame

Usage

get_metadata_names(df)

Arguments

df
dataframe: data to be uploaded to Smartabase

Details

Once data has been pulled in from Smartabase, it is often desirable to retain the metadata variables (e.g. about, start_time etc.) in the data frame before pushing back to Smartabase. Rather than having to repeatedly write out the vector of metadata variables you want to retain, this helper function will retain any metadata variables present in a dataframe; for instance, when used in the select() function

Value

Vector of metadata variable names

Examples

## Not run: 
example_df <- dplyr::tibble(
  about = c("Jamie Anderson", "Charlie Thompson"),
  start_date = c("14/02/2020", "14/02/2020"),
  form = "Hydration",
  `Body Weight pre training` = round(runif(2, 82, 92), 0),
  `Body Weight post training` = round(runif(2, 82, 92), 0),
  `Urine Colour` = round(runif(2, 1, 8), 0),
  end_date = c("14/02/2020", "14/02/2020")
)
example_df %>% select(get_metadata_names(.))
select(example_df, get_metadata_names(example_df))

## End(Not run)

[Package neon version 0.1.20 Index]