neon-package {neon}	R Documentation

neon: R wrapper for Smartabase API

Description

This packages includes functions for pulling data from Smartabase and pushing data to Smartabase. Data pulled into R will be in the form of a data frame and data pushed to Smartabase must be in the form of a data frame (which will be converted into a specific JSON structure required by the Smartabase API).

Author(s)

Maintainer: Zac Pross zac.pross@fusionsport.com

Authors:

James Day james.day@fusionsport.com

[Package neon version 0.1.20 Index]