pull_smartabase {neon}	R Documentation

pull_smartabase

Description

Downloads data from a Smartabase event or profile form

Usage

pull_smartabase(
  form,
  url = NULL,
  type = "event",
  download_attachment = FALSE,
  start_date = NULL,
  end_date = NULL,
  last = NULL,
  start_time = "00:00 AM",
  end_time = "11:59 PM",
  username = NULL,
  password = NULL,
  filter_user_key = NULL,
  filter_user_value = NULL,
  filter_data_key = NULL,
  filter_data_value = NULL,
  filter_data_condition = "equal_to",
  include_missing_users = FALSE,
  guess_col_type = TRUE,
  get_uuid = FALSE,
  cloud_mode = FALSE,
  last_sync_time = NULL,
  shiny_progress_code = NULL
)

Arguments

form
string: name of Smartabase form

url
string: Smartabase url e.g. 'example.smartabase.com/site'

type
string: either 'event', 'profile' or 'synchronise'

download_attachment
boolean: if TRUE, will download any files that are included in the form via a file upload/multiple file upload field

start_date
string: 'dd/mm/yyyy'

end_date
string: 'dd/mm/yyyy'

last
string: string indicating how many of the last days, weeks, months or years data should be pulled for; e.g. '7 days'. Can supply 'today' to pull only today's data. Only works if start_date and end_date are NULL.

start_time
string: 'h:mm AM' or 'h:mm PM'

end_time
string: 'h:mm AM' or 'h:mm PM'

username
string: Smartabase username – ignore if setup with save_credentials()

password
string: Smartabase password – ignore if setup with save_credentials()

filter_user_key
string: 'about', 'username', 'email', 'group' or 'current_group'

filter_user_value
strings: values indicated by 'filter_user_key'. E.g. if filter_user_key = 'about', expects full names of athletes to filter for

filter_data_key
strings: names of Smartabase fields. Multiple can be supplied in a vector

filter_data_value
strings: values indicated by filter_data_key, e.g. if form has a field called 'RPE' and filter_data_key = 'RPE', then valid RPE values must be supplied. Multiple filter_data_value can be supplied in a vector the same length as filter_data_key

filter_data_condition
strings: '=' or 'equal_to'; '!=' or 'not_equal_to'; '%in%' or 'contains'; '<' or 'less_than'; '>' or 'greater_than'; '<=' or 'less_than_or_equal_to'; '>=' or 'greater_than_or_equal_to'. Multiple filter_data_condition can be supplied in a vector the same length as filter_data_key

include_missing_users
boolean: if TRUE, includes users who have not recorded data in that form as rows of NA values

guess_col_type
boolean: if TRUE, pull_smartabase() guesses the the data type of each column and appends the guess to tibble's heading

get_uuid
boolean: if TRUE, return Smartabase uuid

cloud_mode
boolean: if TRUE, no progress bar appears (i.e. for use in non-local, cloud environments)

last_sync_time
numeric: last time pull_smartabase() was synched with server, in epoch time

shiny_progress_code
string: shinyhttr code to pass progress bar to a shiny app

Details

This function pulls data from a Smartabase event form or profile form using the Smartabase API. For more details see the help vignette: vignette("pulling-data", package = "neon")

Value

Smartabase data

Examples

## Not run: 
# Get one week of wellness data from example.smartabase.com/site:
wellness_data <- pull_smartabase(
  url = "example.smartabase.com/site",
  form = "Daily Wellness",
  username = "john.smith",
  password = "examplePassword",
  start_date = "15/04/2019",
  end_date = "22/04/2019"
)

## End(Not run)

[Package neon version 0.1.20 Index]