push_smartabase {neon}	R Documentation

push_smartabase

Description

Uploads a dataframe to a Smartabase event or profile form

Usage

push_smartabase(
  df,
  form,
  url = NULL,
  username = NULL,
  password = NULL,
  entered_by_user_id = NULL,
  type = "event",
  get_id = FALSE,
  match_id_to_column = NULL,
  table_fields = NULL,
  start_date = NULL,
  end_date = NULL,
  current_date_format = NULL,
  start_time = NULL,
  end_time = NULL,
  edit_event = FALSE,
  cloud_mode = FALSE,
  shiny_progress_code = NULL
)

Arguments

df
dataframe: data to be uploaded to Smartabase

form
string: Name of Smartabase form

url
string: Smartabase url e.g. 'example.smartabase.com/site'

username
string: Smartabase username – ignore if setup with save_credentials()

password
string: Smartabase password – ignore if setup with save_credentials()

entered_by_user_id
string: user ID of user uploading the data – ignore if setup with save_credentials()

type
string: either 'event' or 'profile'

get_id
boolean: if TRUE, searches for user IDs of athletes listed in dataframe (requires a value supplied to 'match_id_to_column')

match_id_to_column
string: names of columns in dataframe that match either 'about', 'username' or 'email'. User IDs will be returned that match dataframe values. Requires 'get_id' = TRUE

table_fields
vector: supply a vector of column names that are going to be uploaded into a table. In Smartabase, this is equivalent to ticking 'Treat all records for the same user, on the same day as a single record?'

start_date
string: name of start date column in dataframe. If NULL, and 'Date' or 'start_date' are not already columns in the dataframe, creates a new start_date column based on current date

end_date
string: name of end date column in dataframe. If NULL, and 'end_date' is not already a column in the dataframe, creates a new 'end_date' column based on 'start_date'

current_date_format
string: current format of date variable of interest. Set this argument to convert 'start_date' or 'end_date' to dd/mm/yyyy

start_time
string: name of start time column in dataframe with 'h:mm AM' or 'h:mm PM' records. If NULL, and 'start_time' is not already a column in the dataframe, creates a new 'start_time' column based on Sys.time()

end_time
string: name of end time column in dataframe with 'h:mm AM' or 'h:mm PM' records. If NULL, and 'end_time' is not already a column in the dataframe, creates a new 'end_time' column with values +1 hour after 'start_time'

edit_event
boolean: if TRUE, will look for 'event_id' in dataframe. Records with matching event_id on Smartabase will be overwritten with the new data

cloud_mode
boolean: if TRUE, confirmation pop-up will not appear when editing events (i.e. for use in non-local, cloud environments)

shiny_progress_code
string: shinyhttr code to pass progress bar to a shiny app

Details

The push_smartabase() function uses the Smartabase API and imports and R dataframe into a specific Smartabase event or profile form. Set 'type' = "profile" to push to a profile form. For more details see the help vignette: vignette("getting-started", package = "neon")

Value

http response

Examples

## Not run: 
example_df <- dplyr::tibble(
  user_id = c(31813, 31819),
  `Body Weight pre training` = round(runif(2, 82, 92), 0),
  `Body Weight post training` = round(runif(2, 82, 92), 0),
  `Urine Colour` = round(runif(2, 1, 8), 0)
)
push_smartabase(
  df = example_df,
  url = "example.smartabase.com/site",
  username = "john.smith",
  form = "Hydration",
  entered_by_user_id = 31822
)

## End(Not run)

[Package neon version 0.1.20 Index]