save_credentials {neon}	R Documentation

save_credentials

Description

Allows you to permanently store your Smartabase password, username and user ID on your local machine.

Usage

save_credentials()

Details

Calling this function will open a .Renviron file. This is a plain text hidden file stored on your local machine. Add any of the four variables in the format

SB_USER = "john.smith"
SB_PASS = "examplePassword"
SB_ID = 12345
SB_URL = "example.smartabase.com/site"

...then save the file and restart R. You now don't need to supply the password, username or entered_by_user_id arguments in 'pull_smartabase()' or 'push_smartabase()'. For more details see the help vignette: vignette("getting-started", package = "neon")

Value

Opens .Renviron file

Examples

## Not run: 
save_credentials()

## End(Not run)

[Package neon version 0.1.20 Index]